requirements.txt
pdfplumber
python-docx
openai
faiss-cpu 
numpy
fastapi
uvicorn
python-multipart
streamlit
